Shop-Script v.9.0.5   * Eliminated unnecessary displaying of fractional product quantities if they are disabled in the settings.   * Eliminated export of fractional product quantities to CSV files if they are disabled in the settings.   * Restored the option to save negative product stock quantities.   * Restored the availability of purchasing more product items than are available in stock if this option is enabled in the settings.   * Restore the option to save zero product quantities when editing orders.   * Fixed displaying of negative product quantities in the store backend.     * For the users of the PREMIUM license:     * Fixed saving of selected product quantity units when editing orders.     * Fixed product filtering by their base quantity units’ prices.     * Fixed execution of the order refund action for items with fractional quantities.

+++ lib/workflow/shopWorkflowRefundAction.class.php
@@ -305,7 +305,7 @@ class shopWorkflowRefundAction extends shopWorkflowAction
         $order_items = $order->edit(true, waPayment::OPERATION_REFUND);
         $order_items = $this->workupOrderItems($order, $transaction_data ? $plugin : null, $order_items);
         foreach ($order_items as &$item) {
-            $order_items_count += intval($item['quantity']);
+            $order_items_count += floatval($item['quantity']);
             if ((float)$item['quantity']) {
                 $item['price_with_discount'] = $item['price'] - $item['total_discount'] / $item['quantity'];
             } else {
